+++ webapp/src/service/etc/skillStack.parser.js
